Refactor code for a clean version (#2)
* refactor: Overweighted views and multilines if
* fix: solving issue on editing asset symbol
* fix: adding enum for worker transaction type
* fix on clic functions in line
* Refactor main version to clean code
- Use functions for handlers
- Reduce line amount in big files
- Change logic for multiple conditionals
* fix to clean code deploy
* Reduce props for Contact List component
