+++ frontend/components/Button.tsx
@@ -19,23 +19,29 @@ const button = cva("button", {
       none: ["border-0"],
       underline: ["border-0", "border-b", "!px-0", "!rounded-none", "!py-0"],
     },
+    focusBorder: {
+      none: ["focus-visible:outline-none"],
+      auto: [""],
+    },
   },
   compoundVariants: [
     {
       intent: "noBG",
       border: "underline",
       class: "",
+      focusBorder: "none",
     },
   ],
   defaultVariants: {
     intent: "accept",
     size: "medium",
     border: "none",
+    focusBorder: "none",
   },
 });
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof button> {}
 
-export const CustomButton: FC<ButtonProps> = ({ className, intent, size, border, ...props }) => (
-  <button className={button({ intent, size, border, className })} {...props} />
+export const CustomButton: FC<ButtonProps> = ({ className, intent, size, border, focusBorder, ...props }) => (
+  <button className={button({ intent, size, border, focusBorder, className })} {...props} />
 );
