+++ frontend/pages/hooks/workerHook.tsx
@@ -20,7 +20,12 @@ export const WorkerHook = () => {
           let transactionsICP = await getAllTransactionsICP(elementS.sub_account_id, false);
 
           store.dispatch(
-            setTxWorker({ tx: transactionsICP, symbol: elementA.tokenSymbol, subaccount: elementS.sub_account_id }),
+            setTxWorker({
+              tx: transactionsICP,
+              symbol: elementA.symbol,
+              tokenSymbol: elementA.tokenSymbol,
+              subaccount: elementS.sub_account_id,
+            }),
           );
         });
       } else {
@@ -32,13 +37,17 @@ export const WorkerHook = () => {
               hexToUint8Array(elementS?.sub_account_id || "0x0"),
               false,
               elementA.tokenSymbol,
-              elementA.symbol,
               selectedToken.address,
               elementS?.sub_account_id,
             );
 
             store.dispatch(
-              setTxWorker({ tx: transactionsICRC1, symbol: elementA.symbol, subaccount: elementS.sub_account_id }),
+              setTxWorker({
+                tx: transactionsICRC1,
+                symbol: elementA.symbol,
+                tokenSymbol: elementA.tokenSymbol,
+                subaccount: elementS.sub_account_id,
+              }),
             );
           });
         }
