+++ frontend/redux/assets/AssetActions.tsx
@@ -297,7 +297,6 @@ export const getAllTransactionsICRC1 = async (
   subaccount_index: Uint8Array,
   loading: boolean,
   assetSymbol: string,
-  symbol: string,
   canister: string,
   subNumber?: string,
 ) => {
@@ -319,9 +318,8 @@ export const getAllTransactionsICRC1 = async (
   });
 
   const transactionsInfo = ICRC1getTransactions.transactions.map(({ transaction, id }) =>
-    formatckBTCTransaccion(transaction, id, myPrincipal.toString(), symbol, canister, subNumber),
+    formatckBTCTransaccion(transaction, id, myPrincipal.toString(), assetSymbol, canister, subNumber),
   );
-
   if (
     loading &&
     store.getState().asset.selectedAccount?.sub_account_id === subNumber &&
