+++ frontend/redux/models/AccountModels.ts
@@ -96,8 +96,17 @@ export type RosettaTransaction = z.infer<typeof RosettaTransaction>;
 // Local Storage
 const TransactionList = z.object({
   symbol: z.string(),
+  tokenSymbol: z.string(),
   subaccount: z.string(),
   tx: z.array(Transaction),
 });
 
 export type TransactionList = z.infer<typeof TransactionList>;
+
+// Process Interfaces
+const AssetToAdd = z.object({
+  symbol: z.string(),
+  tokenSymbol: z.string(),
+  logo: z.string().optional(),
+});
+export type AssetToAdd = z.infer<typeof AssetToAdd>;
